Add abstract `play()` method on Tech.
As part of the langauge change, we add an event handler on this.player_.
This causes a bunch of test failures. This is because in the Component
constructor we make sure that this.player_ is set to `this` but only if
a play method is available. This works fine for the Html5 tech or any
tech that implements the play method (all techs) but a lot of our tests
use Tech directly. Adding the abstract method makes those tests pass.