fix: try again on volume feature detection on iOS
On latest iOS, we are seeing times when the volume feature detection is
showing that we are able to change the volume, though, that is not the
case. Instead, on iOS, when we detect that we can control the volume, we
set a short timer to retest and reset the the featuresVolumeControl
property.

Fixes #7040