fix: hide mouse tooltip on touch devices when not scrubbing
Problem: the vjs-user-inactive class triggering the hiding of the
tooltip is added much later on touch devices, therefore the mouse
tooltip stays visible on its last seek target position whenever the user
taps the screen to access the controls.

This change ensures that the mouse tooltip is only visible when the user
is actually scrubbing.