feat: improve sidebar and toc link accessibility
The table of contents to the right of the page
has a list of links that, when hovered,
are given an underline.

There is also a sidebar to the left of the page
with titles and links to reference documentation.

The color contrast on these elements can be
improved by adding a greater color contrast
for visual accessibility.

This commit changes the hover color of these
elements in a visually nice and accessible way.