use responsive units for media queries
if the user changes their browser font size
setting, px unit media queries will not adapt
to these changes and may cause the query to not
take effect, or to take effect at the wrong time.

rem units on the other hand do adjust based on
the browser's font size setting.

this commit switches unit values and types
to enable this behavior.